added exceptions and validation

+++ src/main/java/com/omnia/users/service/UserService.java
@@ -1,13 +1,14 @@
 package com.omnia.users.service;
 
+import com.omnia.users.exceptions.AlreadyExistsException;
+import com.omnia.users.exceptions.NotFoundException;
 import com.omnia.users.model.entities.UserEntity;
 import com.omnia.users.repository.UserRepository;
+import com.omnia.users.util.Validator;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
 import java.util.Collection;
-import java.util.List;
-import java.util.stream.Collectors;
 
 @Service
 public class UserService {
@@ -16,7 +17,12 @@ public class UserService {
     private UserRepository userRepository;
 
     public UserEntity saveUser(UserEntity user) {
-        return userRepository.save(user);
+        Validator.validateUser(user);
+        if (!userRepository.existsByUsername(user.getUsername())) {
+            return userRepository.save(user);
+        } else {
+            throw new AlreadyExistsException("User with username: " + user.getUsername() + " already taken!");
+        }
     }
 
     public Collection<UserEntity> findAll() {
@@ -24,15 +30,22 @@ public class UserService {
     }
 
     public UserEntity findUserById(Long userId) {
-        return userRepository.findById(userId);
+        return userRepository.findById(userId).orElseThrow(() -> new NotFoundException("User with id: " + userId.toString() + " not found"));
     }
 
     public void deleteById(Long userId) {
-        userRepository.deleteById(userId);
+        if (userRepository.existsById(userId)) {
+            userRepository.deleteById(userId);
+        } else {
+            throw new NotFoundException("User with id: " + userId.toString() + " not found");
+        }
     }
 
     public void delete(UserEntity user) {
-        userRepository.delete(user);
+        if (userRepository.existsById(user.getId())) {
+            userRepository.delete(user);
+        } else {
+            throw new NotFoundException("User: " + user + " not found");
+        }
     }
-
 }
